+++ src/bindings/python.rs
@@ -145,6 +145,8 @@ fn completion_item_to_pydict(py: Python<'_>, item: &analyze::CompletionItem) -> 
     dict.set_item("filter_text", item.filter_text.as_str())?;
     dict.set_item("sort_text", item.sort_text.as_str())?;
     dict.set_item("kind", completion_kind_to_str(&item.kind))?;
+    dict.set_item("link_id", item.link_id.as_str())?;
+    dict.set_item("item_type", item.item_type.as_str())?;
     Ok(dict.into())
 }